Added functionality to have different hyperparameters in different models
This is a big commit. Basically, I've moved the hyperparameter
calculations out of MCMC and into the respective models. As of now, the
only hyperparameter each model has is Sphi, but soon Aphi and Sepsilon
will be added. There is a new pure virtual function in Model called
calculateLogLikelihoodForHyperParameters, which is where you guys can do
your accept/reject calculations. This is already done for Sphi however.
Parameter also has a new function called proposeHyperParameters. That
one is self-explanatory.